Version final

+++ front/js/cart.js
@@ -169,79 +169,106 @@ function buildCart() {
     });
   };
 }
-
+///////////////////////////////////////////////////
+//////// CREE DES REGEX POUR CHAQUES LIGNE ////////
+//////////////// DU FORMULAIRE ////////////////////
+///////////// ENVOIE LE FORMULAIRE ////////////////
+///////////////////////////////////////////////////
 function sendConfirmation() {
+  let cart = JSON.parse(localStorage.getItem("produit"));
+  let cartValidation = new Boolean();
+
   const firstName = document.querySelector('#firstName');
   const firstNameErrorMsg = document.querySelector('#firstNameErrorMsg');
+  let firstNameValidation = new Boolean();
+  firstNameValidation = false;
 
   firstName.addEventListener('change', () => {
+    firstNameErrorMsg.textContent = '';
     const firstNameValid = new RegExp(
       '^[A-Z]{1}[a-z]+$', 'g'
     );
     const firstNameTest = firstNameValid.test(firstName.value);
     if(firstNameTest) {
-      firstNameErrorMsg.textContent = '';
+      firstNameValidation = true;
     } else {
+      firstNameValidation = false;
       firstNameErrorMsg.textContent = 'Commence par une majuscule, ne dois contenir ni chiffres, ni accent, ni caractère spéciaux'
     };
   });
 
   const lastName = document.querySelector('#lastName');
-  const lastNameErrorMsg = document.querySelector('#lastNameErrorMsg')
+  const lastNameErrorMsg = document.querySelector('#lastNameErrorMsg');
+  let lastNameValidation = new Boolean();
+  lastNameValidation = false;
 
   lastName.addEventListener('change', () => {
+    lastNameErrorMsg.textContent = '';
     const lastNameValid = new RegExp(
       '^[A-Z]{1}[a-z]+$', 'g'
     );
     const lastNameTest = lastNameValid.test(lastName.value);
     if(lastNameTest) {
-      lastNameErrorMsg.textContent = '';
+      lastNameValidation = true;
     } else {
+      lastNameValidation = false;
       lastNameErrorMsg.textContent = 'Commence par une majuscule, ne dois contenir ni chiffres, ni accent, ni caractère spéciaux';
     };
   });
 
   const address = document.querySelector('#address');
-  const addressErrorMsg = document.querySelector('#addressErrorMsg')
+  const addressErrorMsg = document.querySelector('#addressErrorMsg');
+  let addressValidation = new Boolean();
+  addressValidation = false;
 
   address.addEventListener('change', () => {
+    addressErrorMsg.textContent = '';
     const addressValid = new RegExp(
       '^[0-9]+[a-zA-Z, ]+$', 'g'
     );
     const addressTest = addressValid.test(address.value);
     if(addressTest) {
-      addressErrorMsg.textContent = '';
+      addressValidation = true;
     } else {
+      addressValidation = false;
       addressErrorMsg.textContent = 'Commence par un chiffre puis une suite de lettre';
     };
   });
 
   const city = document.querySelector('#city');
-  const cityErrorMsg = document.querySelector('#cityErrorMsg')
+  const cityErrorMsg = document.querySelector('#cityErrorMsg');
+  let cityValidation = new Boolean();
+  cityValidation = false;
 
   city.addEventListener('change', () => {
+    cityErrorMsg.textContent = '';
     const cityValid = new RegExp(
       '^[A-Za-z- ]+$', 'g'
     );
     const cityTest = cityValid.test(city.value);
     if(cityTest) {
-      cityErrorMsg.textContent = '';
+      cityValidation = true;
     } else {
+      cityValidation = false;
       cityErrorMsg.textContent = 'Ne dois pas contenir de chiffres';
     };
   });
 
   const email = document.querySelector('#email');
-  const emailErrorMsg = document.querySelector('#emailErrorMsg')
+  const emailErrorMsg = document.querySelector('#emailErrorMsg');
+  let emailValidation = new Boolean();
+  emailValidation = false;
 
   email.addEventListener('change', () => {
+    emailErrorMsg.textContent = '';
     const emailValid = new RegExp(
       '^[a-zA-Z0-9.-_]+[@]{1}[a-zA-Z0-9.-_]+[.]{1}[a-z]{2,10}$', 'g'
     );
     const emailTest = emailValid.test(email.value);
     if(emailTest) {
-      emailErrorMsg.textContent = '';
+      emailValidation = true;
     } else {
+      emailValidation = false;
       emailErrorMsg.textContent = 'adress mail doit comporter @ et .';
     };
   });
@@ -249,35 +276,57 @@ function sendConfirmation() {
   const buttonOrder = document.querySelector('#order');
 
   buttonOrder.addEventListener('click', () => {
-    let canapIdArray = [];
-    cart.forEach((canap) => {
-      for(let i = 0; i < canap.nombreCanap; i++) {
-        canapIdArray.push(canap.idCanap);
+
+    if(JSON.parse(localStorage.getItem("produit") === null)) {
+      cartValidation = false;
+    } else {
+      cartValidation = true;
+    }
+
+    if(cartValidation === false) {
+      alert('Votre panier est vide')
+    } else 
+    if(firstNameValidation === false) {
+      alert('Ajoutez un prénom valide');
+    } else if (lastNameValidation === false) {
+      alert('Ajoutez un nom de famille valide');
+    } else if (addressValidation === false) {
+      alert('Ajoutez une adress valide');
+    } else if (cityValidation === false) {
+      alert ('Ajoutez un nom de ville valide');
+    } else if (emailValidation === false) {
+      alert('Ajoutez une adress mail valide');
+    } else {
+      let canapIdArray = [];
+      cart.forEach((canap) => {
+        for(let i = 0; i < canap.nombreCanap; i++) {
+          canapIdArray.push(canap.idCanap);
+        };
+      });
+      let order = {
+        contact: {
+          firstName: firstName.value,
+          lastName: lastName.value,
+          address: address.value,
+          city: city.value,
+          email: email.value,
+        },
+        products: canapIdArray,
       };
-    });
-    let order = {
-      contact: {
-        firstName: firstName.value,
-        lastName: lastName.value,
-        address: address.value,
-        city: city.value,
-        email: email.value,
-      },
-      products: canapIdArray,
-    };
-    const fetchOption = {
-      method: 'POST',
-      body : JSON.stringify(order),
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    };
-    fetch(`http://localhost:3000/api/products/order`, fetchOption).then(async (response) => {
-      order = await response.json();
-      const orderId = order.orderId;
-      localStorage.clear();
-      localStorage.setItem('orderId', orderId);
-      document.location.href = "confirmation.html";
-    });
+      const fetchOption = {
+        method: 'POST',
+        body : JSON.stringify(order),
+        headers: {
+          'Content-Type': 'application/json'
+        }
+      };
+      fetch(`http://localhost:3000/api/products/order`, fetchOption).then(async (response) => {
+        order = await response.json();
+        const orderId = order.orderId;
+        localStorage.clear();
+        localStorage.setItem('orderId', orderId);
+        document.location.href = "confirmation.html" + "?orderID=" + orderId;
+      });
+    }
   });
 }
